move .icc into interface as its used by other packages. Move from nearly empty package into DT package
