+++ RecoLocalMuon/DTRecHit/plugins/BuildFile.xml
@@ -1,6 +1,7 @@
 <use name="FWCore/Framework"/>
 <use name="FWCore/ParameterSet"/>
 <use name="root"/>
+<use name="DataFormats/DTRecHit"/>
 <use name="RecoLocalMuon/DTRecHit"/>
 <use name="MagneticField/Records"/>
 <use name="MagneticField/Engine"/>
