+++ RecoLocalMuon/DTRecHit/plugins/DTDriftTimeParameters.icc
@@ -1,6 +1,6 @@
 #ifndef RecoLocalMuon_DTDriftTimeParameters_H
 #define RecoLocalMuon_DTDriftTimeParameters_H
 
-#include "DataFormats/MuonData/src/DTDriftTimeParameters.icc"
+#include "DataFormats/DTRecHit/interface/DTDriftTimeParameters.icc"
 
 #endif
