+++ SimMuon/DTDigitizer/BuildFile.xml
@@ -1,6 +1,7 @@
 <use name="FWCore/Framework"/>
 <use name="FWCore/ParameterSet"/>
 <use name="DataFormats/Common"/>
+<use name="DataFormats/DTRecHit"/>
 <use name="SimDataFormats/TrackingHit"/>
 <use name="SimDataFormats/CrossingFrame"/>
 <use name="DataFormats/DTDigi"/>
