+++ SimMuon/DTDigitizer/test/BuildFile.xml
@@ -11,6 +11,7 @@
   <use name="FWCore/Framework"/>
   <use name="FWCore/ParameterSet"/>
   <use name="DataFormats/MuonDetId"/>
+  <use name="DataFormats/DTRecHit"/>
   <use name="Geometry/DTGeometry"/>
   <use name="Geometry/Records"/>
   <use name="DataFormats/DTDigi"/>
